refactor: moved civic center service to its own module
